third_party/blink/web_tests/svg/animations/change-baseVal-while-animating-fill-freeze-expected.txt

SVG 1.1 dynamic animation tests

This tests scripting baseVal while animation is running

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rect.width.animVal.value is 10
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 30
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 30
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 50
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 50
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 50
PASS rect.width.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
